requirements.txt
scipy==1.2.1
numpy==1.17.0
matplotlib==3.1.2
opencv_python==4.1.2.30
torch==1.2.0
torchvision==0.4.0
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0
# 注意： 还需要编译DCN模块，DCN模块源码目录为./nets/DCNv2_latest-master.zip , 将./nets下原始的DCNv2删除，解压DCNv2_latest-master.zip到当前目录，并将文件名称改为DCNv2，在所用的虚拟环境下对DCNv2进行编译即可。